+++ frontend/src/app/app.component.html
@@ -1,8 +1,5 @@
 <app-header></app-header>
 <div class="content">
-    <select #langSelect (change)="translate.use(langSelect.value)">
-        <option *ngFor="let lang of translate.getLangs()" [value]="lang" [selected]="lang === translate.currentLang">{{ lang }}</option>
-      </select>
   <router-outlet></router-outlet>
 </div>
 <app-footer></app-footer>
