+++ frontend/src/app/core/header/header.component.html
@@ -10,6 +10,9 @@
           <i class="icon icon__pnone-number"></i>
           <p class="header-contacts__text">453 - 5553 - 996</p>
         </a>
+        <select #langSelect (change)="translate.use(langSelect.value)">
+            <option *ngFor="let lang of translate.getLangs()" [value]="lang" [selected]="lang === translate.currentLang">{{ lang }}</option>
+        </select>
       </div>
       <div class="header-contacts__social-media">
         <a href="#" target="_blank" class="icon icon__facebook"></a>
